pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "spatialfusion"
description = "SpatialFusion: Deep learning models for spatial omics data analysis."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "Yates, Josephine", email = "yatesjos@broadinstitute.org" }]
keywords = ["spatial omics", "deep learning", "bioinformatics", "multi-omics"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering :: Bio-Informatics"
]

# Keep accelerator stacks out of here; provide extras below instead.
dependencies = [
  "numpy>=1.24,<3.0",
  "pandas>=2.0,<3.0",
  "scipy>=1.10,<2.0",
  "pyarrow>=12",
  "tqdm>=4.65",
  "hydra-core>=1.3",
  "omegaconf>=2.3",
  "scanpy>=1.9,<2.0",
  "anndata>=0.10,<0.12",          
  "matplotlib>=3.7",              
  "Pillow>=9.5",                  
  "tifffile>=2023.7.10",
  "scikit-learn>=1.2",
  "seaborn>=0.12",                
  "shapely>=2.0",
  "tensorboard>=2.12",
  "imagecodecs>=2023.1.23",
  "networkx>=3.2",
  "h5py>=3.8"                     
]

dynamic = ["version"]

[project.optional-dependencies]
# Choose one of these during install; do NOT include in default deps.
cpu = [
  "torch>=2.3,<2.6",
  "torchvision>=0.18,<0.21"
]
docs = [
  "sphinx>=7.0",
  "myst-parser>=2.0",
  "sphinx-rtd-theme>=2.0"
]
dev = [
  "pytest>=8",
  "pytest-cov>=5",
  "ruff>=0.5",
  "black>=24.0",
  "pre-commit>=3.6"
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.setuptools_scm]
version_scheme = "post-release"
local_scheme = "node-and-date"
fallback_version = "0.1.0"

[project.urls]
Homepage = "https://github.com/your-org/spatialfusion"
Documentation = "https://readthedocs.org/projects/your-docs"
Issues = "https://github.com/your-org/spatialfusion/issues"